Ready for use

+++ protected/model/Passenger.php
@@ -28,20 +28,16 @@ class Passenger
     }
 
     public static function delete_by_ride($ride_id, $conn){
-        $query = "DELETE FROM `passenger` WHERE `ride_id` = $ride_id)";
+        $query = "DELETE FROM `passenger` WHERE `ride_id` = $ride_id";
         $res = mysqli_query($conn, $query);
-        print $query;
         return $res ? true : false;
     }
 
     public static function delete_by_user_and_ride($user_id, $ride_id, $conn){
-        $query = "DELETE FROM `passenger` WHERE `user_id` = $user_id AND `ride_id` = $ride_id)";
+        $query = "DELETE FROM `passenger` WHERE `user_id` = $user_id AND `ride_id` = $ride_id";
         $res = mysqli_query($conn, $query);
         return $res ? true : false;
     }
 
-    private static function row_to_object($row){
-
-    }
 
 }
